Do not restart browser when test type changes
At least for Chrome, parameters used to start Chrome won't change when
test type changes (except for wdspec tests).

This change tries to switch to a different executor where allowed, and
reuse the webdriver session from the previous executor.

A browser side flag is used to control whether this behavior should be
enabled or not. It is up to each browser vendor to decide whether it
should be enabled or not.

Bug: 341581718